requirements.txt
django >= 3.1.*
django-inline-svg >= 0.1.*
requests >= 2.*
psycopg2 >= 2.8.*           ; platform_system=='Windows'
psycopg2-binary  >= 2.8.*   ; platform_system=='Linux'
django-heroku >= 0.3.*
gunicorn >= 20.1.*
whitenoise >= 5.2.*
dj-database-url >= 0.5.*
sentry-sdk >= 1.0.*
python-dotenv >= 0.17.*
newrelic
